+++ components/EDebriefSurvey.vue
@@ -1,26 +1,3 @@
-<script lang="ts" setup>
-import type { SurveyPage } from '~/core/components/epochs/ESurveySequence.vue'
-
-const pages: SurveyPage[] = [
-  {
-    id: 'difficulty',
-    prompt: 'How difficult was the experiment?',
-    options: ['too easy', 'just right', 'too hard'],
-  },
-  {
-    id: 'fun',
-    prompt: 'How fun was the experiment?',
-    options: ['worse than average', 'typical', 'better than average'],
-  },
-  {
-    id: 'feedback',
-    kind: 'text',
-    prompt: 'Do you have any other feedback?',
-    placeholder: 'Optional',
-  },
-]
-</script>
-
 <template>
   <div w-600px mx-auto>
     <h1>You're done!</h1>
@@ -29,6 +6,22 @@ const pages: SurveyPage[] = [
       Thanks for participating! We have a few quick questions before you go.
     </p>
 
-    <ESurveySequence name="Survey" :pages="pages" />
+    <ESurveyWrapper name="Survey">
+      <ESurveyButtons
+        name="difficulty"
+        prompt="How difficult was the experiment?"
+        options="too easy, just right, too hard"
+      />
+      <ESurveyButtons
+        name="fun"
+        prompt="How fun was the experiment?"
+        options="worse than average, typical, better than average"
+      />
+      <ESurveyText
+        name="feedback"
+        prompt="Do you have any other feedback?"
+        placeholder="Optional"
+      />
+    </ESurveyWrapper>
   </div>
 </template>